feat(io_uring): implement Slab<User> (#2252)

+++ examples/rust/src/multi-tenant/producer/main.rs
@@ -331,7 +331,7 @@ async fn create_stream_and_user(
     info!("Created stream: {stream_name} with ID: {}", stream.id);
     let mut streams_permissions = AHashMap::new();
     streams_permissions.insert(
-        stream.id,
+        stream.id as usize,
         StreamPermissions {
             read_stream: true,
             manage_topics: true,
